ref(ui): delete container components
