+++ frontend/src/components/Nav.tsx
@@ -9,8 +9,8 @@ import {
 } from "@/components/Dropdown"
 import { Chevron } from "@/components/icons"
 import Logo from "@/components/Logo"
+import { NavLink } from "@/components/Routing"
 import { UserDropdown } from "@/components/UserDropdown"
-import NavLink from "@/containers/NavLink"
 import { useDispatch, useSelector } from "@/hooks"
 import { scheduleURLFromTeamID } from "@/store/mapState"
 import { fetchingUserAsync } from "@/store/thunks"
