+++ frontend/src/components/Notification.tsx
@@ -1,10 +1,17 @@
-import { INotificationState } from "@/store/reducers/notification"
+import { connect } from "react-redux"
+
+import {
+  clearNotification,
+  INotificationState,
+} from "@/store/reducers/notification"
+import { IState } from "@/store/store"
+import { Dispatch } from "@/store/thunks"
 
 interface INotificationProps extends INotificationState {
   readonly close: () => void
 }
 
-const notification = ({
+const Notification = ({
   message,
   level = "info",
   show = true,
@@ -30,4 +37,19 @@ const notification = ({
   return null
 }
 
-export default notification
+const mapDispatchToProps = (dispatch: Dispatch) => {
+  return {
+    close: () => dispatch(clearNotification()),
+  }
+}
+
+const mapStateToProps = (state: IState) => {
+  return {
+    message: state.notification.message,
+    closeable: state.notification.closeable,
+    show: state.notification.show,
+    level: state.notification.level,
+  }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Notification)
