+++ frontend/src/components/Routing.tsx
@@ -1,4 +1,9 @@
-import { Link as RRLink, LinkProps as RRLinkProps } from "react-router-dom"
+import {
+  Link as RRLink,
+  LinkProps as RRLinkProps,
+  NavLinkProps,
+  useLocation,
+} from "react-router-dom"
 
 interface ILinkProps extends RRLinkProps {
   readonly isRaw?: boolean
@@ -10,4 +15,29 @@ export function Link({ to, replace, isRaw, ...rest }: ILinkProps) {
     return <a href={to} {...rest} />
   }
   return <RRLink to={to} replace={replace} {...rest} />
+}
+
+export const NavLink = ({
+  to,
+  className = "",
+  activeClassName = "active",
+  onClick,
+  children,
+}: {
+  className: string
+  activeClassName?: string
+  to: NavLinkProps["to"]
+  onClick?: () => void
+  children?: React.ReactNode
+}) => {
+  const location = useLocation()
+  const activeClass = location.pathname === to ? activeClassName : ""
+  return (
+    <RRLink
+      to={to}
+      className={className + " " + activeClass}
+      onClick={onClick}
+      children={children}
+    />
+  )
 }
